Display duration and reward
