+++ packages/ui/src/working-groups/types/index.ts
@@ -1,1 +1,2 @@
 export * from './WorkingGroup'
+export * from './WorkingGroupOpening'
